requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# CUDA and compilation
ninja>=1.11.0

# Testing
pytest>=7.4.0
hypothesis>=6.82.0
pytest-cov>=4.1.0

# Profiling and benchmarking
nvidia-ml-py3>=7.352.0
psutil>=5.9.0

# Utilities
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
